% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{getFormats}
\alias{getFormats}
\title{List available FORMAT fields in a VCFReader object}
\arguments{
\item{x}{A `VCFReader` object.}
}
\value{
A character vector of FORMAT field names.
}
\description{
Returns the names of FORMAT fields available in the VCF header (e.g., "GT", "DS", "GP").
}
\examples{
filename <- system.file("extdata", "example.vcf.gz", package = "readVCF")
a <- openVCF(filename)
getFormats(a)

}
\seealso{
\link{VCFReader}
}
