% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{getLine}
\alias{getLine}
\alias{VCFnext}
\title{read a line in a VCF file}
\arguments{
\item{x}{description}

\item{field}{}
}
\value{
a list with components values, CHR ...
}
\description{
Read current line in a readVCF object
}
\details{
blabla (GT et DS)
}
\examples{
a <- openVCF(filename, "19:320000-700000")
sample.ids <- getSamples(a)
geno <- matrix(nrow = 0, ncol = length(sample.ids), dimnames = list(character(0), sample.ids))
while( VCFnext(a) ) {
  line <- getLine(a, "DS")
  geno <- rbind(geno, line$values)
}

}
