% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{getSamples}
\alias{getSamples}
\title{Get sample IDs from a VCFReader object}
\arguments{
\item{x}{A `VCFReader` object.}
}
\value{
A character vector of sample IDs.
}
\description{
Extracts the list of sample identifiers from the header of a VCF file opened with [openVCF()].
}
\examples{
filename <- system.file("extdata", "example.vcf.gz", package = "readVCF")
a <- openVCF(filename)
getSamples(a)

}
