% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openVCF.r
\name{openVCF}
\alias{openVCF}
\title{Opening VCF files}
\usage{
openVCF(filename, regions)
}
\arguments{
\item{filename}{name of file to open}

\item{regions}{(optional) regions to read}
}
\value{
an object of class VCFReader
}
\description{
Creating an object to read VCF files line by line
}
\details{
blabla
}
\examples{
filename <- system.file("extdata", "example.vcf.gz", package = "readVCF")
a <- openVCF(filename)
getFormats(a)
}
\seealso{
[getFormats()]
}
