% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.VCFreader.r
\name{print.VCFReader}
\alias{print.VCFReader}
\title{Print a VCFReader object}
\usage{
\method{print}{VCFReader}(x, ...)
}
\arguments{
\item{x}{An object of class `VCFReader`.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
No return value. This function is invoked for its side effect (printing to the console).
}
\description{
Prints a short summary of a `VCFReader` object, including the filename and available FORMAT fields.
}
\examples{
filename <- system.file("extdata", "example.vcf.gz", package = "readVCF")
reader <- openVCF(filename)
print(reader)  # or just `reader`


}
\seealso{
[openVCF()]
}
