% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{VCFnext}
\alias{VCFnext}
\title{Move to the next line in a VCFReader object}
\arguments{
\item{x}{A `VCFReader` object.}
}
\value{
`TRUE` if a new line was successfully read, `FALSE` if end of file is reached.
}
\description{
Advances the internal pointer to the next line in the VCF file.
Should be used in a loop to iterate over lines.
}
\examples{
filename <- system.file("extdata", "example.vcf.gz", package = "readVCF")
a <- openVCF(filename)
VCFnext(a)

}
\seealso{
\link{getLine}, \link{VCFReader}
}
